refactor: predict shop stuff
