+++ Content.Server/_Scp/Scp106/Systems/Scp106System.cs
@@ -2,11 +2,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Content.Server._Scp.Helpers;
 using Content.Server._Scp.Scp106.Components;
-using Content.Server.Actions;
 using Content.Server.DoAfter;
 using Content.Server.GameTicking;
 using Content.Server.Gateway.Systems;
-using Content.Server.Popups;
 using Content.Server.Store.Systems;
 using Content.Server.Stunnable;
 using Content.Shared._Scp.Scp106;
@@ -16,7 +14,6 @@ using Content.Shared.Alert;
 using Content.Shared.Body.Components;
 using Content.Shared.DoAfter;
 using Content.Shared.FixedPoint;
-using Content.Shared.Hands.Components;
 using Content.Shared.Hands.EntitySystems;
 using Content.Shared.Humanoid;
 using Content.Shared.Interaction;
@@ -24,7 +21,6 @@ using Content.Shared.Mind;
 using Content.Shared.Mobs;
 using Content.Shared.Mobs.Components;
 using Content.Shared.Mobs.Systems;
-using Content.Shared.Store.Components;
 using Robust.Server.Audio;
 using Robust.Shared.Audio;
 using Robust.Shared.Map;
@@ -46,10 +42,8 @@ public sealed partial class Scp106System : SharedScp106System
     [Dependency] private readonly AlertsSystem _alerts = default!;
     [Dependency] private readonly StunSystem _stun = default!;
     [Dependency] private readonly DoAfterSystem _doAfter = default!;
-    [Dependency] private readonly ActionsSystem _actions = default!;
     [Dependency] private readonly MobStateSystem _mobState = default!;
     [Dependency] private readonly SharedHandsSystem _hands = default!;
-    [Dependency] private readonly PopupSystem _popup = default!;
     [Dependency] private readonly GameTicker _gameTicker = default!;
     [Dependency] private readonly IRobustRandom _random = default!;
     [Dependency] private readonly IGameTiming _timing = default!;
@@ -75,31 +69,18 @@ public sealed partial class Scp106System : SharedScp106System
 
         SubscribeLocalEvent((Entity<Scp106BackRoomMapComponent> _, ref AttemptGatewayOpenEvent args) => args.Cancelled = true);
 
-        #region Phantom
-
-        SubscribeLocalEvent<Scp106PhantomComponent, MobStateChangedEvent>(OnMobStateChangedEvent);
-        SubscribeLocalEvent<Scp106PhantomComponent, Scp106ReverseActionEvent>(OnScp106ReverseActionEvent);
-
-        #endregion
+        SubscribeLocalEvent<Scp106PortalSpawnerComponent, ComponentInit>(OnPortalSpawn);
 
         #region Store & Its abilities
 
         SubscribeLocalEvent<Scp106Component, Scp106ShopAction>(OnShop);
 
-        // Abilities in that store - I love lambdas >:)
+        #endregion
 
-        // TODO: Проверка на хендхелд и кенселед
-        SubscribeLocalEvent((Entity<Scp106Component> ent, ref Scp106BoughtPhantomAction args) =>
-            _actions.AddAction(ent, args.BoughtAction));
-        SubscribeLocalEvent((Entity<Scp106Component> ent, ref Scp106OnUpgradePhantomAction args) =>
-            ent.Comp.PhantomCoolDown -= args.CooldownReduce);
-        SubscribeLocalEvent((Entity<Scp106Component> ent, ref Scp106BoughtBareBladeAction args) =>
-            _actions.AddAction(ent, args.BoughtAction));
-        SubscribeLocalEvent((Entity<Scp106Component> ent, ref Scp106BoughtCreatePortal args) =>
-            _actions.AddAction(ent, args.BoughtAction));
+        #region Phantom
 
-        SubscribeLocalEvent<Scp106Component, Scp106BareBladeAction>(OnScp106BareBladeAction);
-        SubscribeLocalEvent<Scp106PortalSpawnerComponent, ComponentInit>(OnPortalSpawn);
+        SubscribeLocalEvent<Scp106PhantomComponent, MobStateChangedEvent>(OnMobStateChangedEvent);
+        SubscribeLocalEvent<Scp106PhantomComponent, Scp106ReverseActionEvent>(OnScp106ReverseActionEvent);
 
         #endregion
 
@@ -305,73 +286,6 @@ public sealed partial class Scp106System : SharedScp106System
 
     #endregion
 
-    #region Shop & Its abilities
-
-    private void OnShop(EntityUid uid, Scp106Component component, Scp106ShopAction args)
-    {
-        if (!TryComp<StoreComponent>(uid, out var store))
-            return;
-
-        _store.ToggleUi(uid, uid, store);
-    }
-
-    private void OnScp106BareBladeAction(Entity<Scp106Component> ent, ref Scp106BareBladeAction args)
-    {
-        TryToggleBlade(ent, ref args);
-    }
-
-    private bool TryToggleBlade(Entity<Scp106Component> ent, ref Scp106BareBladeAction args, bool force = false)
-    {
-        if (args.Handled)
-            return false;
-
-        // Клинок можно использовать только в карманном измерении или форсированно через код
-        if (!CheckIsInDimension(ent) && !force)
-        {
-            var message = Loc.GetString("scp106-bare-blade-action-invalid");
-            _popup.PopupEntity(message, ent, ent);
-
-            return false;
-        }
-
-        ToggleBlade(ent, args.Prototype);
-
-        args.Handled = true;
-        return true;
-    }
-
-    private void ToggleBlade(Entity<Scp106Component> ent, EntProtoId blade)
-    {
-        // Если клинок уже имеется
-        if (ent.Comp.HandTransformed)
-        {
-            HideBlade(ent);
-        }
-        else
-        {
-            EnsureComp<HandsComponent>(ent);
-            _hands.AddHand(ent, "right", HandLocation.Middle);
-            var sword = Spawn(blade, Transform(ent).Coordinates);
-
-            ent.Comp.Sword = sword;
-            _hands.TryPickup(ent, sword, "right");
-
-            ent.Comp.HandTransformed = true;
-        }
-    }
-
-    public void HideBlade(Entity<Scp106Component> ent)
-    {
-        if (!Exists(ent.Comp.Sword))
-            return;
-
-        Del(ent.Comp.Sword);
-        _hands.RemoveHands(ent);
-        ent.Comp.HandTransformed = false;
-    }
-
-    #endregion
-
     #region Helpers
 
     private async Task<EntityCoordinates> GetTransferMark()
@@ -387,7 +301,7 @@ public sealed partial class Scp106System : SharedScp106System
 
     private HashSet<EntityCoordinates> SearchForMarks()
     {
-        return EntityQuery<Scp106BackRoomMarkComponent>()
+        return _scpHelpers.GetAll<Scp106BackRoomMarkComponent>()
             .Select(entity => Transform(entity.Owner).Coordinates)
             .ToHashSet();
     }
